Fix compilation warning.

+++ shared/async/src/windows/async.c
@@ -461,7 +461,7 @@ static int write_callback(int device) {
     }
 
     if(devices[device].callback.fp_write) {
-        devices[device].callback.fp_write(devices[device].callback.user, ret == 0 ? dwBytesWritten : -1);
+        devices[device].callback.fp_write(devices[device].callback.user, ret == 0 ? (int)dwBytesWritten : -1);
     }
 
     return ret;
